Fix out-of-bounds error determining current day
There is a bug in the timing plugin that causes it to always return the
"closed" path on Sundays, regardless of the time. This is because the
calculation of the index into the array of weekdays goes to -1 on
Sundays. The patch fixes this by correctly wrapping the index around,
modulus-style.